refactor: cmake and interfaces
Retooled the cmake configuration such that the hardware library
is now an INTERFACE library, rather than directly modifying the
root project. This allows us to remove numerous link_lib defs
from the cli and root project, as well as provides an easier method
to pass hardware-level compiler definitions up to the main project
in the cases where that is necessary.