20 wifi support - Initial Work (#22)
* refactor: cmake and interfaces

Retooled the cmake configuration such that the hardware library
is now an INTERFACE library, rather than directly modifying the
root project. This allows us to remove numerous link_lib defs
from the cli and root project, as well as provides an easier method
to pass hardware-level compiler definitions up to the main project
in the cases where that is necessary.

* feat(#20): need configUSE_PASSIVE_IDLE_HOOK defined for SMP versions of FreeRTOS

* feat(#20): more necessary FreeRTOS config updates

* feat(#20): include lwip_sys_freertos lib, decrease heap space accordingly

* wip: wifi functions

* wip: implemented all necessary wifi wrapper func

* wip: correctly toggle hw_wifi.c inclusion based on HW_WIFI

* wip: first pass at the actual service

* fix(#20): apparently, configNUM_CORES still has to be defined!?!?!?

* fix(#20): watchdog priority, update cli printing from wifi

* chore: update with mcknly's PR comments